+++ core/__init__.py
@@ -5,6 +5,6 @@ SST Lounge Bot - Core Package
 from .bot import SSTLoungeBot
 from .database import SimpleDB
 
-__version__ = "1.5.1"
+__version__ = "1.5.2"
 __author__ = "Unofficial SST Team"
 __all__ = ["SSTLoungeBot", "SimpleDB"]